[front/home] category/search/main route
