+++ front/public/js/home/draw_page/drawNotFound.js
@@ -1,0 +1,5 @@
+export default function drawNotFound() {
+  const body = document.querySelector('body');
+
+  body.innerHTML = 'go back';
+}
