+++ front/views/home/home.html
@@ -13,7 +13,7 @@
     <div id="wrap">
       <div id="header">
         <h1 id="logo">
-          <a href="/home"
+          <a href="/home#main/"
             ><img src="/img/home/Logo.png" style="width: 160px; height: 80px"
           /></a>
         </h1>
